refactor: refactoring vr contents hook, types
